pyproject.toml
[project]
name = "legado"
version = "0.1.0"
description = "A Legado book source synchronization tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.2",
    "markdown2>=2.5.1",
    "pathspec>=0.12.1",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
]

[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
dev-dependencies = [
    "pre-commit>=4.0.1",
    "ruff>=0.7.4",
]

[tool.ruff]
line-length = 88
indent-width = 4
fix = true
show-fixes = true

exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "migrations",
    "web",
]


[tool.ruff.format]
docstring-code-format = true
indent-style = "space"
line-ending = "lf"
quote-style = "double"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "N",
    "UP",
    "ASYNC",
    "B",
    "A",
    "SLF",
    "SIM",
    "I",
    "PIE",
    "ARG",
    "PTH",
    "TD",
    "FIX",
    "Q",
    "RSE",
    "RET",
    "TID",
    "TCH",
    "PGH",
    "PL",
    "FLY",
    "FURB",
    "RUF",
]
ignore = [
    "E501",
    "E701",
    "F541",
    "W291",
]

[tool.ruff.lint.isort]
case-sensitive = true
combine-as-imports = true
split-on-trailing-comma = true
force-wrap-aliases = true
order-by-type = true
